WIP: background task for syncronizing NAT information for service zones
