+++ nexus/src/app/background/mod.rs
@@ -14,6 +14,7 @@ mod inventory_collection;
 mod nat_cleanup;
 mod phantom_disks;
 mod status;
+mod sync_service_zone_nat;
 
 pub use common::Driver;
 pub use common::TaskHandle;
