+++ nexus/src/app/background/sync_service_zone_nat.rs
@@ -1,0 +1,6 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Background task for detecting changes to service zone locations and
+//! updating the NAT rpw table accordingly
